Separate private api tests from public api tests
